apps/api/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.32,<1.0
sqlalchemy[asyncio]>=2.0,<3.0
aiomysql>=0.2,<1.0
pymysql>=1.1,<2.0
cryptography>=42,<44
alembic>=1.13,<2.0
pydantic>=2.8,<3.0
pydantic-settings>=2.6,<3.0
email-validator>=2.2,<3.0
passlib[argon2]>=1.7,<2.0
argon2-cffi>=23,<24
python-jose[cryptography]>=3.3,<4.0
python-multipart>=0.0.17,<1.0
